CRA-337 Made class cravatrend.cpp holding the trend cubes and the trend cube sampling. All trends are sampled from min to max of its corresponding trend cube, using 1000 sample points.
The class contains the functions CravaTrend::getTrendPosition() giving the trend parameters s1 and s2 used to find trend values, and CravaTrend::GetTrendCubeSampling() giving the sampling of the trends, used when the trends needs to be resampled.

git-svn-id: https://code.nr.no/scm/svn/crava/branches/StatoilDevelopment@1393 176de4d3-4747-0410-aa96-47f6d6634214

+++ rplib/rockphysicsstorage.cpp
@@ -51,10 +51,10 @@ GaussianRockPhysicsStorage::~GaussianRockPhysicsStorage()
 }
 
 DistributionsRock *
-GaussianRockPhysicsStorage::GenerateRockPhysics(const std::string                      & path,
-                                                const std::vector<std::string>         & trend_cube_parameters,
-                                                const std::vector<std::vector<float> > & trend_cube_sampling,
-                                                std::string                            & errTxt) const
+GaussianRockPhysicsStorage::GenerateRockPhysics(const std::string                       & path,
+                                                const std::vector<std::string>          & trend_cube_parameters,
+                                                const std::vector<std::vector<double> > & trend_cube_sampling,
+                                                std::string                             & errTxt) const
 {
   NRLib::Trend * mean_vp_trend                = mean_vp_               ->GenerateTrend(path,trend_cube_parameters,trend_cube_sampling,errTxt);
   NRLib::Trend * mean_vs_trend                = mean_vs_               ->GenerateTrend(path,trend_cube_parameters,trend_cube_sampling,errTxt);
